Move transformations into their own module operations
